Updates to vdrv (vctl driver subcommand) and installation tooling.
